run eslint
