+++ package.json
@@ -25,6 +25,7 @@
     "@tauri-apps/plugin-log": "^2.0.0-beta.6",
     "@tauri-apps/plugin-os": "^2.0.0-beta.5",
     "@tauri-apps/plugin-process": "^2.0.0-beta.5",
+    "@tauri-apps/plugin-shell": "^2.0.0-beta.6",
     "@tauri-apps/plugin-updater": "^2.0.0-beta.5",
     "ag-grid-react": "^31.3.2",
     "apexcharts": "^3.49.1",
@@ -33,7 +34,7 @@
     "highcharts-react-official": "^3.2.1",
     "immer": "^10.1.1",
     "non.geist": "^1.0.3",
-    "posthog-js": "^1.136.5",
+    "posthog-js": "^1.139.3",
     "react": "^18.3.1",
     "react-apexcharts": "^1.4.1",
     "react-arborist": "^3.4.0",
@@ -49,24 +50,24 @@
     "zustand": "^4.5.2"
   },
   "devDependencies": {
-    "@eslint/js": "^9.4.0",
+    "@eslint/js": "^9.5.0",
     "@tauri-apps/cli": "next",
     "@types/lodash.sortby": "^4.7.9",
-    "@types/node": "^20.14.0",
+    "@types/node": "^20.14.6",
     "@types/react": "^18.3.3",
     "@types/react-dom": "^18.3.0",
-    "@vitejs/plugin-react": "^4.3.0",
+    "@vitejs/plugin-react": "^4.3.1",
     "@vitejs/plugin-react-swc": "^3.7.0",
-    "eslint": "8.x",
+    "eslint": "^8.57.0",
     "eslint-config-react-app": "^7.0.1",
     "eslint-plugin-no-relative-import-paths": "^1.5.4",
-    "eslint-plugin-react": "^7.34.2",
+    "eslint-plugin-react": "^7.34.3",
     "eslint-plugin-unused-imports": "^4.0.0",
-    "globals": "^15.4.0",
+    "globals": "^15.6.0",
     "husky": "^9.0.11",
     "typescript": "^5.4.5",
-    "typescript-eslint": "^7.12.0",
-    "vite": "^5.2.12",
+    "typescript-eslint": "^7.13.1",
+    "vite": "^5.3.1",
     "vite-plugin-eslint": "^1.8.1"
   },
   "packageManager": "yarn@3.8.2"
