+++ src-tauri/gen/schemas/capabilities.json
@@ -1,1 +1,1 @@
-{"main":{"identifier":"main","description":"permissions for pachtop","local":true,"windows":["main"],"permissions":["store:allow-get","store:allow-set","store:allow-save","os:allow-arch","os:allow-hostname","os:allow-os-type","os:allow-platform","window:allow-start-dragging","autostart:allow-is-enabled","updater:allow-check","updater:default","log:allow-log","log:default"]},"migrated":{"identifier":"migrated","description":"permissions that were migrated from v1","local":true,"windows":["main"],"permissions":["path:default","event:default","window:default","app:default","resources:default","menu:default","tray:default"]}}
+{"main":{"identifier":"main","description":"permissions for pachtop","local":true,"windows":["main"],"permissions":["store:allow-get","store:allow-set","store:allow-save","os:allow-arch","os:allow-hostname","os:allow-os-type","os:allow-platform","window:allow-start-dragging","autostart:allow-is-enabled","updater:allow-check","updater:default","log:allow-log","log:default","shell:allow-spawn"]},"migrated":{"identifier":"migrated","description":"permissions that were migrated from v1","local":true,"windows":["main"],"permissions":["path:default","event:default","window:default","app:default","resources:default","menu:default","tray:default"]}}
