+++ src-tauri/tauri.conf.json
@@ -45,6 +45,15 @@
   "version": "1.0.0-1",
   "identifier": "com.pachtop.dev",
   "plugins": {
+    "shell": {
+      "scope": [
+        {
+          "name": "powershell",
+          "cmd": "powershell",
+          "args": true
+        }
+      ]
+    },
     "updater": {
       "pubkey": "dW50cnVzdGVkIGNvbW1lbnQ6IG1pbmlzaWduIHB1YmxpYyBrZXk6IEM0NkIyODA1NThCM0I0MUEKUldRYXRMTllCU2hyeE9mSUx0MElJN2N5VFo5TytFV1RwaUtoaytDNVdmajNQMmlseVlYMGYzeUkK",
       "endpoints": [
